requirements.txt
pandas==2.1.1
websockets>=12.0
python-dotenv>=1.0.0